@@ -35,15 +35,15 @@
     "@angular/platform-server": "^12.2.13",
     "@angular/router": "^12.2.13",
     "@elastic/apm-rum-angular": "^2.1.1",
-    "@fortawesome/angular-fontawesome": "^0.8.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-regular-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
+    "@fortawesome/angular-fontawesome": "0.9.0",
+    "@fortawesome/fontawesome-svg-core": "1.2.36",
+    "@fortawesome/free-brands-svg-icons": "5.15.4",
+    "@fortawesome/free-regular-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@gorniv/ngx-transfer-http": "^2.2.9",
     "@nguniversal/common": "^12.1.3",
     "@nguniversal/express-engine": "^12.1.3",
-    "angular-auth-oidc-client": "^11.6.4",
+    "angular-auth-oidc-client": "12.0.3",
     "array-flat-polyfill": "^1.0.1",
     "body-parser": "^1.19.0",
     "bootstrap": "^4.6.0",
@@ -80,6 +80,11 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~12.2.13",
+    "@angular-eslint/builder": "12.7.0",
+    "@angular-eslint/eslint-plugin": "12.7.0",
+    "@angular-eslint/eslint-plugin-template": "12.7.0",
+    "@angular-eslint/schematics": "12.7.0",
+    "@angular-eslint/template-parser": "12.7.0",
     "@angular/cli": "^12.2.13",
     "@angular/compiler-cli": "^12.2.13",
     "@angular/language-service": "^12.2.13",
@@ -95,10 +100,14 @@
     "@types/lodash": "^4.14.168",
     "@types/lodash-es": "^4.17.5",
     "@types/node": "^14.14.35",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
+    "@typescript-eslint/eslint-plugin": "5.7.0",
+    "@typescript-eslint/parser": "5.7.0",
     "codelyzer": "^6.0.1",
-    "eslint": "^7.26.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-typescript": "^16.1.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-prettier": "^4.0.0",
     "https-proxy-agent": "^5.0.0",
     "jasmine-core": "^3.7.1",
     "jasmine-spec-reporter": "~6.0.0",
@@ -107,6 +116,7 @@
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
+    "prettier": "^2.5.1",
     "protractor": "~7.0.0",
     "ts-loader": "^8.0.18",
     "ts-node": "~9.1.1",
