@@ -35,14 +35,14 @@ export class AuthGuard implements CanActivate {
     // if (this.appSettingsService.myDataSettings.develop) return of(true);
 
     return this.oidcSecurityService.isAuthenticated$.pipe(
-      map((isAuthorized: boolean) => {
-        if (!isAuthorized) {
+      map((authResult: any) => {
+        if (!authResult?.isAuthenticated) {
           this.router.navigate(['/mydata']);
           return false;
         }
 
         return true;
-      })
+      }),
     );
   }
 }
