@@ -32,7 +32,7 @@ export class CollaborationsService {
   }
 
   updateTokenInHttpAuthHeader() {
-    const token = this.oidcSecurityService.getToken();
+    const token = this.oidcSecurityService.getAccessToken();
 
     this.httpOptions = {
       headers: new HttpHeaders({
