@@ -38,7 +38,7 @@ export class FundingsService {
   }
 
   updateTokenInHttpAuthHeader() {
-    const token = this.oidcSecurityService.getToken();
+    const token = this.oidcSecurityService.getAccessToken();
 
     this.httpOptions = {
       headers: new HttpHeaders({
