@@ -37,7 +37,7 @@ export class ProfileService {
   }
 
   updateTokenInHttpAuthHeader() {
-    const token = this.oidcSecurityService.getToken();
+    const token = this.oidcSecurityService.getAccessToken();
 
     this.httpOptions = {
       headers: new HttpHeaders({
